Create camera base backend for generic cameras
Allows specific camera implementations to be created easier by having
this handle all the other camera logic besides image capture (like
telemetry and gimbal things in the future).
